+++ slack-app-backend/src/Configuration/AppConfiguration.cs
@@ -1,0 +1,11 @@
+
+using System;
+
+namespace SlackAppBackend.Configuration
+{
+    public class AppConfiguration : IAppConfiguration
+    {
+        public Uri SystemEventServiceUri => new Uri(Environment.GetEnvironmentVariable("SYSTEM_EVENTS_SVC_URI"));
+        public string SlackBotUserOAuthAccessToken => Environment.GetEnvironmentVariable("SLACK_OAUTHACCESS_TOKEN");
+    }
+}