+++ slack-app-backend/src/Configuration/IAppConfiguration.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace SlackAppBackend.Configuration
+{
+    public interface IAppConfiguration
+    {
+        Uri SystemEventServiceUri { get; }
+
+        string SlackBotUserOAuthAccessToken { get; }
+    }
+}