+++ slack-app-backend/src/Enums/SlackCommand.cs
@@ -1,4 +1,4 @@
-namespace SystemEvents.Enums
+namespace SlackAppBackend.Enums
 {
     public enum SlackCommand
     {