+++ slack-app-backend/src/Program.cs
@@ -11,7 +11,7 @@ namespace SystemEvents
 {
     public class Program
     {
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
+        private static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(webBuilder => {
                     webBuilder.UseStartup<Startup>()