+++ system-events/src/Configuration/ISlackApiConfiguration.cs
@@ -1,7 +1,0 @@
-namespace SystemEvents.Configuration
-{
-    public interface ISlackApiConfiguration
-    {
-        public string SlackBotUserOAuthAccessToken { get; }
-    }
-}